# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  coins: number;
  spins: number;
  totalReferrals: number;
  createdAt: string;
}

export interface GlobalStats {
  totalUsers: number;
  totalCoinsIssued: number;
  totalSpins: number;
  totalRedemptions: number;
}

export type WheelSegmentRewardType = typeof WheelSegmentRewardType[keyof typeof WheelSegmentRewardType];


export const WheelSegmentRewardType = {
  coins: 'coins',
  spins: 'spins',
  gift: 'gift',
} as const;

export interface WheelSegment {
  id: number;
  label: string;
  rewardType: WheelSegmentRewardType;
  rewardAmount: number;
  /** @nullable */
  giftId?: number | null;
  /** Probability weight (relative, not necessarily summing to 1) */
  probability: number;
  color: string;
  enabled: boolean;
}

export type WheelSegmentInputRewardType = typeof WheelSegmentInputRewardType[keyof typeof WheelSegmentInputRewardType];


export const WheelSegmentInputRewardType = {
  coins: 'coins',
  spins: 'spins',
  gift: 'gift',
} as const;

export interface WheelSegmentInput {
  label: string;
  rewardType: WheelSegmentInputRewardType;
  rewardAmount: number;
  /** @nullable */
  giftId?: number | null;
  probability: number;
  color: string;
  enabled?: boolean;
}

export type WheelSegmentUpdateRewardType = typeof WheelSegmentUpdateRewardType[keyof typeof WheelSegmentUpdateRewardType];


export const WheelSegmentUpdateRewardType = {
  coins: 'coins',
  spins: 'spins',
  gift: 'gift',
} as const;

export interface WheelSegmentUpdate {
  label?: string;
  rewardType?: WheelSegmentUpdateRewardType;
  rewardAmount?: number;
  /** @nullable */
  giftId?: number | null;
  probability?: number;
  color?: string;
  enabled?: boolean;
}

export interface SpinRequest {
  userId: number;
}

export interface SpinResult {
  segmentId: number;
  label: string;
  rewardType: string;
  rewardAmount: number;
  newCoins: number;
  newSpins: number;
  landingAngle?: number;
}

export interface SpinRecord {
  id: number;
  userId: number;
  segmentId: number;
  label: string;
  rewardType: string;
  rewardAmount: number;
  createdAt: string;
}

export interface ReferralEntry {
  id: number;
  firstName: string;
  /** @nullable */
  username: string | null;
  joinedAt: string;
}

export interface ReferralInfo {
  userId: number;
  referralLink: string;
  totalReferrals: number;
  referralsRequired: number;
  rewardType: string;
  rewardAmount: number;
  nextRewardAt: number;
  referralList: ReferralEntry[];
}

export interface Gift {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  coinPrice: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  stock?: number | null;
  enabled: boolean;
}

export interface GiftInput {
  name: string;
  description?: string;
  coinPrice: number;
  imageUrl?: string;
  /** @nullable */
  stock?: number | null;
  enabled?: boolean;
}

export interface GiftUpdate {
  name?: string;
  description?: string;
  coinPrice?: number;
  imageUrl?: string;
  /** @nullable */
  stock?: number | null;
  enabled?: boolean;
}

export interface RedeemRequest {
  userId: number;
}

export type RedemptionStatus = typeof RedemptionStatus[keyof typeof RedemptionStatus];


export const RedemptionStatus = {
  pending: 'pending',
  fulfilled: 'fulfilled',
  cancelled: 'cancelled',
} as const;

export interface Redemption {
  id: number;
  userId: number;
  giftId: number;
  giftName: string;
  /** @nullable */
  giftImageUrl?: string | null;
  coinsCost: number;
  status: RedemptionStatus;
  createdAt: string;
}

export type AppSettingsReferralRewardType = typeof AppSettingsReferralRewardType[keyof typeof AppSettingsReferralRewardType];


export const AppSettingsReferralRewardType = {
  coins: 'coins',
  spins: 'spins',
} as const;

export interface AppSettings {
  maintenanceMode: boolean;
  wheelEnabled: boolean;
  referralsEnabled: boolean;
  giftsEnabled: boolean;
  redeemEnabled: boolean;
  accountEnabled: boolean;
  referralsRequired: number;
  referralRewardType: AppSettingsReferralRewardType;
  referralRewardAmount: number;
  botUsername: string;
}

export type SettingsUpdateReferralRewardType = typeof SettingsUpdateReferralRewardType[keyof typeof SettingsUpdateReferralRewardType];


export const SettingsUpdateReferralRewardType = {
  coins: 'coins',
  spins: 'spins',
} as const;

export interface SettingsUpdate {
  maintenanceMode?: boolean;
  wheelEnabled?: boolean;
  referralsEnabled?: boolean;
  giftsEnabled?: boolean;
  redeemEnabled?: boolean;
  accountEnabled?: boolean;
  referralsRequired?: number;
  referralRewardType?: SettingsUpdateReferralRewardType;
  referralRewardAmount?: number;
  botUsername?: string;
}

export interface AdminVerifyInput {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface AdminUsersPage {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type BalanceAdjustmentType = typeof BalanceAdjustmentType[keyof typeof BalanceAdjustmentType];


export const BalanceAdjustmentType = {
  coins: 'coins',
  spins: 'spins',
} as const;

export interface BalanceAdjustment {
  type: BalanceAdjustmentType;
  /** Positive to add, negative to remove */
  amount: number;
}

export interface AdminLog {
  id: number;
  action: string;
  /** @nullable */
  details: string | null;
  createdAt: string;
}

export type GetMeParams = {
telegramId: string;
firstName?: string;
lastName?: string;
username?: string;
referralCode?: string;
};

export type GetSpinHistoryParams = {
userId: number;
limit?: number;
};

export type GetReferralsParams = {
userId: number;
};

export type GetMyRedemptionsParams = {
userId: number;
};

export type AdminGetUsersParams = {
search?: string;
page?: number;
limit?: number;
};

export type AdminGetLogsParams = {
limit?: number;
};