# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 7DOGS Telegram Mini App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/verify:
    post:
      operationId: adminVerify
      tags: [admin]
      summary: Verify admin password and get a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyInput"
      responses:
        "200":
          description: Token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Invalid password
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users (paginated, searchable)
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{userId}/balance:
    patch:
      operationId: adminAdjustBalance
      tags: [admin]
      summary: Add or remove coins/spins for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceAdjustment"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/wheel:
    get:
      operationId: adminGetWheel
      tags: [admin]
      summary: Get all wheel segments (including disabled)
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WheelSegment"
    post:
      operationId: adminCreateSegment
      tags: [admin]
      summary: Add a wheel segment
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WheelSegmentInput"
      responses:
        "201":
          description: Created segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WheelSegment"
  /admin/wheel/{segmentId}:
    patch:
      operationId: adminUpdateSegment
      tags: [admin]
      parameters:
        - in: path
          name: segmentId
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WheelSegmentUpdate"
      responses:
        "200":
          description: Updated segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WheelSegment"
    delete:
      operationId: adminDeleteSegment
      tags: [admin]
      parameters:
        - in: path
          name: segmentId
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/gifts:
    get:
      operationId: adminGetGifts
      tags: [admin]
      summary: Get all gifts (including disabled) for admin management
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gift"
    post:
      operationId: adminCreateGift
      tags: [admin]
      summary: Add a new gift
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftInput"
      responses:
        "201":
          description: Created gift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gift"
  /admin/gifts/{giftId}:
    patch:
      operationId: adminUpdateGift
      tags: [admin]
      parameters:
        - in: path
          name: giftId
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftUpdate"
      responses:
        "200":
          description: Updated gift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gift"
    delete:
      operationId: adminDeleteGift
      tags: [admin]
      parameters:
        - in: path
          name: giftId
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/settings:
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update app settings (maintenance, section disables, referral rules)
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
  /admin/logs:
    get:
      operationId: adminGetLogs
      tags: [admin]
      summary: Get admin action logs
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLog"
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, firstName, coins, spins, totalReferrals, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        coins:
          type: integer
        spins:
          type: integer
        totalReferrals:
          type: integer
        createdAt:
          type: string
...
    WheelSegment:
      type: object
      required: [id, label, rewardType, rewardAmount, probability, color, enabled]
      properties:
        id:
          type: integer
        label:
          type: string
        rewardType:
          type: string
          enum: [coins, spins, gift]
        rewardAmount:
          type: integer
        giftId:
          type: ["integer", "null"]
        probability:
          type: number
          description: Probability weight (relative, not necessarily summing to 1)
        color:
          type: string
        enabled:
          type: boolean
    WheelSegmentInput:
      type: object
      required: [label, rewardType, rewardAmount, probability, color]
      properties:
        label:
          type: string
        rewardType:
          type: string
          enum: [coins, spins, gift]
        rewardAmount:
          type: integer
        giftId:
          type: ["integer", "null"]
        probability:
          type: number
        color:
          type: string
        enabled:
          type: boolean
          default: true
    WheelSegmentUpdate:
      type: object
      properties:
        label:
          type: string
        rewardType:
          type: string
          enum: [coins, spins, gift]
        rewardAmount:
          type: integer
        giftId:
          type: ["integer", "null"]
        probability:
          type: number
        color:
          type: string
        enabled:
          type: boolean
...
    Gift:
      type: object
      required: [id, name, coinPrice, enabled, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coinPrice:
          type: integer
        imageUrl:
          type: ["string", "null"]
        stock:
          type: ["integer", "null"]
        enabled:
          type: boolean
    GiftInput:
      type: object
      required: [name, coinPrice]
      properties:
        name:
          type: string
        description:
          type: string
        coinPrice:
          type: integer
        imageUrl:
          type: string
        stock:
          type: ["integer", "null"]
        enabled:
          type: boolean
          default: true
    GiftUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coinPrice:
          type: integer
        imageUrl:
          type: string
        stock:
          type: ["integer", "null"]
        enabled:
          type: boolean
...
    AppSettings:
      type: object
      required: [maintenanceMode, wheelEnabled, referralsEnabled, giftsEnabled, redeemEnabled, accountEnabled, referralsRequired, referralRewardType, referralRewardAmount, botUsername]
      properties:
        maintenanceMode:
          type: boolean
        wheelEnabled:
          type: boolean
        referralsEnabled:
          type: boolean
        giftsEnabled:
          type: boolean
        redeemEnabled:
          type: boolean
        accountEnabled:
          type: boolean
        referralsRequired:
          type: integer
        referralRewardType:
          type: string
          enum: [coins, spins]
        referralRewardAmount:
          type: integer
        botUsername:
          type: string
    SettingsUpdate:
      type: object
      properties:
        maintenanceMode:
          type: boolean
        wheelEnabled:
          type: boolean
        referralsEnabled:
          type: boolean
        giftsEnabled:
          type: boolean
        redeemEnabled:
          type: boolean
        accountEnabled:
          type: boolean
        referralsRequired:
          type: integer
        referralRewardType:
          type: string
          enum: [coins, spins]
        referralRewardAmount:
          type: integer
        botUsername:
          type: string
    AdminVerifyInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminToken:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AdminUsersPage:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BalanceAdjustment:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [coins, spins]
        amount:
          type: integer
          description: Positive to add, negative to remove
    AdminLog:
      type: object
      required: [id, action, details, createdAt]
      properties:
        id:
          type: integer
        action:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string