# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
import type { AppSettingsReferralRewardType } from './appSettingsReferralRewardType';

export interface AppSettings {
  maintenanceMode: boolean;
  wheelEnabled: boolean;
  referralsEnabled: boolean;
  giftsEnabled: boolean;
  redeemEnabled: boolean;
  accountEnabled: boolean;
  referralsRequired: number;
  referralRewardType: AppSettingsReferralRewardType;
  referralRewardAmount: number;
  botUsername: string;
}