# lib/api-zod/src/generated/types/balanceAdjustment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
import type { BalanceAdjustmentType } from './balanceAdjustmentType';

export interface BalanceAdjustment {
  type: BalanceAdjustmentType;
  /** Positive to add, negative to remove */
  amount: number;
}