# lib/api-zod/src/generated/types/gift.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface Gift {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  coinPrice: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  stock?: number | null;
  enabled: boolean;
}