# lib/api-zod/src/generated/types/giftUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftUpdate {
  name?: string;
  description?: string;
  coinPrice?: number;
  imageUrl?: string;
  /** @nullable */
  stock?: number | null;
  enabled?: boolean;
}