# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7DOGS Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  coins: number;
  spins: number;
  totalReferrals: number;
  createdAt: string;
}